Split utils into atomic files for maintainability
